Added pseudoops .byte and .word to assembler so raw data an be defined. Though this, vectors can be defined directly in assembly so the vector defining routines in romloader.c become obsolete. Additionaly the simple byte-for-byte program output in the assembler has been changed to display disassembled code

+++ src/disassembler.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-uint8_t disassemble_line(char **output_line, uint8_t *memory, uint16_t pc, bool current);
+uint8_t disassemble_line(char **output_line, uint8_t *memory, uint32_t pc, bool current);
 
 #endif
